chore: agops inter reserve add -> agops reserve add
 - chore: bid "want" is required but different from proposal.want
