+++ packages/agoric-cli/test/snapshots/test-inter-cli.js.md
@@ -21,7 +21,6 @@ Generated by [AVA](https://avajs.dev).
     Commands:␊
       liquidation                       liquidation commands␊
       bid                               auction bidding commands␊
-      reserve                           reserve commands␊
       vbank                             vbank asset commands␊
       help [command]                    display help for command`
 
@@ -117,20 +116,6 @@ Generated by [AVA](https://avajs.dev).
       --from <address>  wallet address literal or name␊
       --generate-only   print wallet action only␊
       -h, --help        display help for command`
-
-## Usage: inter reserve add
-
-> Command usage:
-
-    `Usage: agops inter reserve add [options]␊
-    ␊
-    add collateral to the reserve␊
-    ␊
-    Options:␊
-      --give <number>              Collateral to give␊
-      --collateral-brand <string>  Collateral brand key (default: "IbcATOM")␊
-      --offer-id <string>          Offer id (default: "addCollateral-978307200000")␊
-      -h, --help                   display help for command`
 
 ## Usage: inter vbank list
