+++ packages/agoric-cli/test/test-inter-cli.js
@@ -372,7 +372,6 @@ usageTest('inter bid by-price');
 usageTest('inter bid by-discount');
 usageTest('inter bid list');
 usageTest('inter bid cancel');
-usageTest('inter reserve add');
 usageTest('inter vbank list');
 
 test('formatBid', t => {
